fix(condo): DOMA-11661 fix build
