+++ apps/condo/domains/ai/components/AIInputNotification.tsx
@@ -5,7 +5,7 @@ import { CheckCircle, Close, RefreshCw, XCircle } from '@open-condo/icons'
 import { useIntl } from '@open-condo/next/intl'
 import { Space, Tooltip, Typography, Button } from '@open-condo/ui'
 
-import { ProgressLoader } from '@condo/domains/common/components/Loader'
+import { ProgressLoader } from '@condo/domains/common/components/ProgressLoader'
 
 import styles from './AIInputNotification.module.css'
