+++ apps/condo/domains/common/components/Loader.tsx
@@ -3,29 +3,7 @@ import styled from '@emotion/styled'
 import { Spin, SpinProps } from 'antd'
 import React from 'react'
 
-import styles from './Loader.module.css'
-
 import { colors } from '../constants/style'
-
-export const ProgressLoader = () => {
-    return (
-        <svg xmlns='http://www.w3.org/2000/svg' width='24' height='24' viewBox='0 0 56 56' fill='none' className={styles.loaderIcon}>
-            <path fillRule='evenodd' clipRule='evenodd'
-                d='M28 8C25.3736 8 22.7728 8.51731 20.3463 9.52241C17.9198 10.5275 15.715 12.0007 13.8579 13.8579C12.0007 15.715 10.5275 17.9198 9.52241 20.3463C8.51732 22.7728 8 25.3736 8 28C8 30.6264 8.51731 33.2272 9.52241 35.6537C10.5275 38.0802 12.0007 40.285 13.8579 42.1421C15.715 43.9993 17.9198 45.4725 20.3463 46.4776C22.7728 47.4827 25.3736 48 28 48C30.6264 48 33.2272 47.4827 35.6537 46.4776C38.0802 45.4725 40.285 43.9993 42.1421 42.1421C43.9993 40.285 45.4725 38.0802 46.4776 35.6537C47.4827 33.2272 48 30.6264 48 28C48 25.3736 47.4827 22.7728 46.4776 20.3463C45.4725 17.9198 43.9993 15.715 42.1421 13.8579C40.285 12.0007 38.0802 10.5275 35.6537 9.52241C33.2272 8.51732 30.6264 8 28 8ZM17.2849 2.13137C20.682 0.724241 24.323 -4.76837e-07 28 0C31.677 4.76837e-07 35.318 0.724242 38.7151 2.13137C42.1123 3.53851 45.199 5.60097 47.799 8.20101C50.399 10.8011 52.4615 13.8878 53.8686 17.2849C55.2758 20.682 56 24.323 56 28C56 31.677 55.2758 35.318 53.8686 38.7151C52.4615 42.1123 50.399 45.199 47.799 47.799C45.1989 50.399 42.1123 52.4615 38.7151 53.8686C35.318 55.2758 31.677 56 28 56C24.323 56 20.682 55.2758 17.2849 53.8686C13.8877 52.4615 10.801 50.399 8.20101 47.799C5.60097 45.1989 3.5385 42.1123 2.13137 38.7151C0.724241 35.318 -4.76837e-07 31.677 0 28C2.38419e-07 24.323 0.724242 20.682 2.13137 17.2849C3.53851 13.8877 5.60097 10.801 8.20101 8.20101C10.8011 5.60097 13.8878 3.5385 17.2849 2.13137Z'
-                fill='#E1E5ED'/>
-            <path fillRule='evenodd' clipRule='evenodd'
-                d='M52 24C54.2091 24 56 25.7909 56 28C56 33.913 54.1281 39.6743 50.6525 44.458C47.1769 49.2417 42.2761 52.8024 36.6525 54.6296C31.0289 56.4568 24.9711 56.4568 19.3475 54.6296C13.7239 52.8024 8.82311 49.2417 5.34753 44.458C4.04903 42.6708 4.44522 40.1693 6.23246 38.8708C8.01969 37.5723 10.5212 37.9685 11.8197 39.7557C14.3022 43.1727 17.8028 45.716 21.8197 47.0211C25.8365 48.3263 30.1635 48.3263 34.1803 47.0211C38.1972 45.716 41.6978 43.1727 44.1803 39.7557C46.6629 36.3388 48 32.2236 48 28C48 25.7909 49.7909 24 52 24Z'
-                fill='url(#paint0_linear_9982_2495)'/>
-            <defs>
-                <linearGradient id='paint0_linear_9982_2495' x1='4.58331' y1='40' x2='56' y2='40'
-                    gradientUnits='userSpaceOnUse'>
-                    <stop stopColor='#26C756'/>
-                    <stop offset='1' stopColor='#4BA2E4'/>
-                </linearGradient>
-            </defs>
-        </svg>
-    )
-}
 
 const DEFAULT_DELAY = 200 // milliseconds
